+++ ui/src/app/components/apps-panel/expanded-app.spec.tsx
@@ -25,7 +25,11 @@ import { GKE_APP_PROXY_PATH_SUFFIX } from 'app/utils/constants';
 import { currentWorkspaceStore } from 'app/utils/navigation';
 import { canDeleteStatuses } from 'app/utils/runtime-utils';
 import { runtimeStore, serverConfigStore } from 'app/utils/stores';
-import { appTypeToString } from 'app/utils/user-apps-utils';
+import {
+  appTypeToString,
+  toAppType,
+  UIAppType,
+} from 'app/utils/user-apps-utils';
 import {
   AppsApi as LeoAppsApi,
   ProxyApi,
@@ -55,7 +59,6 @@ import {
 } from 'testing/utils';
 
 import { ExpandedApp } from './expanded-app';
-import { toAppType, UIAppType } from './utils';
 
 const googleProject = 'project-for-test';
 const workspace = {
