+++ ui/src/app/components/apps-panel/pause-resume-button.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { useEffect, useState } from 'react';
 import { faPause, faPlay, faSyncAlt } from '@fortawesome/free-solid-svg-icons';
 
+import { UserEnvironmentStatus } from '../common-env-conf-panels/user-environment-status';
 import { cond, switchCase } from '@terra-ui-packages/core-utils';
 
 import { AppsPanelButton } from './apps-panel-button';
-import { UserEnvironmentStatus } from './utils';
 
 interface Props {
   externalStatus: UserEnvironmentStatus;
