+++ ui/src/app/components/breadcrumb.spec.tsx
@@ -1,5 +1,6 @@
 import { WorkspacesApi } from 'generated/fetch';
 
+import { UIAppType } from '../utils/user-apps-utils';
 import { getTrail } from 'app/components/breadcrumb';
 import {
   analysisTabName,
@@ -16,7 +17,6 @@ import { ConceptSetsApiStub } from 'testing/stubs/concept-sets-api-stub';
 import { workspaceDataStub } from 'testing/stubs/workspaces';
 import { WorkspacesApiStub } from 'testing/stubs/workspaces-api-stub';
 
-import { UIAppType } from './apps-panel/utils';
 import { BreadcrumbType } from './breadcrumb-type';
 
 describe('getTrail', () => {
