+++ ui/src/app/components/common-env-conf-panels/start-stop-environment-button.spec.tsx
@@ -4,7 +4,7 @@ import { AppStatus, RuntimeStatus } from 'generated/fetch';
 
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { UIAppType } from 'app/components/apps-panel/utils';
+import { UIAppType } from 'app/utils/user-apps-utils';
 
 import {
   StartStopEnvironmentButton,
