+++ ui/src/app/components/gke-app-config-panel-wrapper.tsx
@@ -7,9 +7,13 @@ import { switchCase } from '@terra-ui-packages/core-utils';
 import { Spinner } from 'app/components/spinners';
 import { appsApi, disksApi } from 'app/services/swagger-fetch-clients';
 import { notificationStore } from 'app/utils/stores';
-import { deleteUserApp, findDisk } from 'app/utils/user-apps-utils';
+import {
+  deleteUserApp,
+  findApp,
+  findDisk,
+  toUIAppType,
+} from 'app/utils/user-apps-utils';
 
-import { findApp, toUIAppType } from './apps-panel/utils';
 import { ConfirmDelete } from './common-env-conf-panels/confirm-delete';
 import { ConfirmDeleteEnvironmentWithPD } from './common-env-conf-panels/confirm-delete-environment-with-pd';
 import { ConfirmDeleteUnattachedPD } from './common-env-conf-panels/confirm-delete-unattached-pd';
