+++ ui/src/app/components/gke-app-configuration-panels/create-gke-app-button.tsx
@@ -9,13 +9,16 @@ import {
 } from 'generated/fetch';
 
 import { cond } from '@terra-ui-packages/core-utils';
-import { isAppActive } from 'app/components/apps-panel/utils';
 import { Button } from 'app/components/buttons';
 import { TooltipTrigger } from 'app/components/popups';
 import { SUPPORT_EMAIL } from 'app/components/support';
 import { ApiErrorResponse, fetchWithErrorModal } from 'app/utils/errors';
 import { NotificationStore } from 'app/utils/stores';
-import { appTypeToString, createUserApp } from 'app/utils/user-apps-utils';
+import {
+  appTypeToString,
+  createUserApp,
+  isAppActive,
+} from 'app/utils/user-apps-utils';
 
 export interface CreateGkeAppButtonProps {
   createAppRequest: CreateAppRequest;
