+++ ui/src/app/components/gke-app-configuration-panels/gke-app-config-panel-main.tsx
@@ -10,13 +10,6 @@ import {
 } from 'generated/fetch';
 
 import { switchCase } from '@terra-ui-packages/core-utils';
-import {
-  canDeleteApp,
-  defaultCromwellConfig,
-  defaultRStudioConfig,
-  defaultSASConfig,
-  toUIAppType,
-} from 'app/components/apps-panel/utils';
 import { LinkButton } from 'app/components/buttons';
 import { DeletePersistentDiskButton } from 'app/components/common-env-conf-panels/delete-persistent-disk-button';
 import { EnvironmentInformedActionPanel } from 'app/components/common-env-conf-panels/environment-informed-action-panel';
@@ -29,7 +22,12 @@ import { sidebarActiveIconStore } from 'app/utils/navigation';
 import { ProfileStore } from 'app/utils/stores';
 import {
   appTypeToString,
+  canDeleteApp,
+  defaultCromwellConfig,
+  defaultRStudioConfig,
+  defaultSASConfig,
   isInteractiveUserApp,
+  toUIAppType,
   unattachedDiskExists,
 } from 'app/utils/user-apps-utils';
 import { WorkspaceData } from 'app/utils/workspace-data';
