+++ ui/src/app/components/gke-app-configuration-panels/open-gke-app-button.spec.tsx
@@ -8,9 +8,9 @@ import {
   UserAppEnvironment,
 } from 'generated/fetch';
 
+import { UIAppType } from '../../utils/user-apps-utils';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { UIAppType } from 'app/components/apps-panel/utils';
 import { appDisplayPath } from 'app/routing/utils';
 import { registerApiClient } from 'app/services/swagger-fetch-clients';
