+++ ui/src/app/components/help-sidebar.tsx
@@ -14,9 +14,9 @@ import {
   RuntimeStatus,
 } from 'generated/fetch';
 
+import { UIAppType } from '../utils/user-apps-utils';
 import { switchCase } from '@terra-ui-packages/core-utils';
 import { AppsPanel } from 'app/components/apps-panel';
-import { UIAppType } from 'app/components/apps-panel/utils';
 import { CloseButton, StyledExternalLink } from 'app/components/buttons';
 import { ConfigurationPanel } from 'app/components/configuration-panel';
 import { ConfirmWorkspaceDeleteModal } from 'app/components/confirm-workspace-delete-modal';
