+++ ui/src/app/components/runtime-configuration-panel/create-panel.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { Profile, RuntimeStatus, Workspace } from 'generated/fetch';
 
-import { UIAppType } from 'app/components/apps-panel/utils';
+import { UIAppType } from '../../utils/user-apps-utils';
 import { Button } from 'app/components/buttons';
 import { EnvironmentInformedActionPanel } from 'app/components/common-env-conf-panels/environment-informed-action-panel';
 import { styles } from 'app/components/common-env-conf-panels/styles';
