+++ ui/src/app/pages/admin/workspace/disks-table.tsx
@@ -4,7 +4,7 @@ import * as fp from 'lodash/fp';
 import { Column } from 'primereact/column';
 import { DataTable } from 'primereact/datatable';
 
-import { UIAppType } from 'app/components/apps-panel/utils';
+import { UIAppType } from '../../../utils/user-apps-utils';
 import { Button } from 'app/components/buttons';
 import { Spinner } from 'app/components/spinners';
 import { disksAdminApi } from 'app/services/swagger-fetch-clients';
